+++ example/lib/example5_themes_playground/widgets/panels/widget_showcase/widget_showcase.dart
@@ -1,8 +1,5 @@
 import 'package:flutter/material.dart';
 
-import '../../../../shared/widgets/app/show_color_scheme_colors.dart';
-import '../../../../shared/widgets/app/show_sub_theme_colors.dart';
-import '../../../../shared/widgets/app/show_theme_data_colors.dart';
 import '../../../../shared/widgets/universal/showcase_material.dart';
 
 /// All the widgets using effective theming in a column.
@@ -17,12 +14,13 @@ class WidgetShowcase extends StatelessWidget {
         crossAxisAlignment: CrossAxisAlignment.start,
         children: <Widget>[
           SizedBox(height: 8),
-          ShowColorSchemeColors(),
-          SizedBox(height: 8),
-          ShowThemeDataColors(),
-          SizedBox(height: 8),
-          ShowSubThemeColors(),
-          SizedBox(height: 16),
+          // TODO(rydmike): Removed colors, they are under effective colors.
+          // ShowColorSchemeColors(),
+          // SizedBox(height: 8),
+          // ShowThemeDataColors(),
+          // SizedBox(height: 8),
+          // ShowSubThemeColors(),
+          // SizedBox(height: 16),
           ShowcaseMaterial(),
           SizedBox(height: 8),
         ],